[Provider] clean up annotation use by making it non null.
